requirements.txt
streamlit==1.32.0
langchain==0.1.12
openai==0.28.1
google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pymongo[srv]==3.11
python-dotenv==1.0.1
celery==5.3.6
fastapi==0.110.0
uvicorn==0.27.1
pandas==2.0.3
numpy==1.24.3
pytube==15.0.0
moviepy==1.0.3
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1 